Consolidate setup of a new ColonyNetwork into helper function
